Implement new schemas for banx staking
