+++ src/api/common/staking/requests.ts
@@ -1,50 +1,33 @@
 import axios from 'axios'
 import { BN } from 'fbonds-core'
 
-import {
-  BanxStakeInfoResponse,
-  BanxStakeInfoResponseSchema,
-  BanxStakingSettings,
-  BanxStakingSettingsSchema,
-} from '@banx/api/common/staking/schemas'
 import { parseResponseSafe } from '@banx/api/shared'
 import { BACKEND_BASE_URL, BANX_TOKEN_DECIMALS } from '@banx/constants'
 import { ZERO_BN } from '@banx/utils'
 
-import { convertToBanxInfoBN, convertToBanxStakingSettingsBN } from './converters'
-import { BanxInfoBN, BanxStakingSettingsBN } from './types'
+import { BanxStakingInfoSchema, BanxStakingSettingsSchema } from './schemas'
+import { BanxStakingInfo, BanxStakingSettings } from './types'
 
-type FetchBanxStakeInfo = (props: { userPubkey?: string }) => Promise<BanxInfoBN | null>
+type FetchBanxStakeInfo = (props: { userPubkey?: string }) => Promise<BanxStakingInfo | undefined>
 export const fetchBanxStakeInfo: FetchBanxStakeInfo = async ({ userPubkey }) => {
   const queryParams = new URLSearchParams({
     walletPubkey: userPubkey ?? '',
   })
 
-  const { data } = await axios.get<{ data: BanxStakeInfoResponse }>(
+  const { data } = await axios.get<{ data: BanxStakingInfo }>(
     `${BACKEND_BASE_URL}/staking/v2/info?${queryParams.toString()}`,
   )
 
-  if (!data?.data) return null
-
-  const parsedData = (await parseResponseSafe(
-    data.data,
-    BanxStakeInfoResponseSchema,
-  )) as BanxStakeInfoResponse
-
-  return convertToBanxInfoBN(parsedData)
+  return await parseResponseSafe<BanxStakingInfo>(data.data, BanxStakingInfoSchema)
 }
 
-type FetchBanxStakeSettings = () => Promise<BanxStakingSettingsBN | null>
+type FetchBanxStakeSettings = () => Promise<BanxStakingSettings | undefined>
 export const fetchBanxStakeSettings: FetchBanxStakeSettings = async () => {
   const { data } = await axios.get<{ data: BanxStakingSettings }>(
     `${BACKEND_BASE_URL}/staking/v2/settings`,
   )
 
-  if (!data?.data) return null
-
-  await parseResponseSafe(data.data, BanxStakingSettingsSchema)
-
-  return convertToBanxStakingSettingsBN(data.data)
+  return await parseResponseSafe<BanxStakingSettings>(data.data, BanxStakingSettingsSchema)
 }
 
 export const fetchBanxTokenCirculatingAmount = async () => {
